+++ modulo-4/mobileAutomation/browserstack.yml
