+++ modulo-4/mobileAutomation/src/main/java/pages/WikipediaPage.java
@@ -13,7 +13,6 @@ import java.time.Duration;
 import java.util.List;
 
 public class WikipediaPage {
-    private final AppiumDriver driver;
     private final WebDriverWait wait;
 
     @FindBy(id = "org.wikipedia.alpha:id/fragment_onboarding_skip_button")
@@ -23,7 +22,6 @@ public class WikipediaPage {
     public WebElement insertTextElement;
 
     public WikipediaPage(AppiumDriver driver) {
-        this.driver = driver;
         this.wait = new WebDriverWait(driver, Duration.ofSeconds(30));
         PageFactory.initElements(driver, this);
     }
