+++ modulo-4/mobileAutomation/src/test/java/tests/WikipediaTest.java
@@ -1,20 +1,25 @@
 package tests;
 
+import io.appium.java_client.android.AndroidDriver;
+import io.appium.java_client.android.options.UiAutomator2Options;
+import org.openqa.selenium.MutableCapabilities;
+import org.testng.Assert;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 import pages.WikipediaPage;
-import pages.YouTubePage;
 
 import java.net.MalformedURLException;
+import java.net.URL;
 
 public class WikipediaTest extends BaseTest {
     private WikipediaPage wikipediaPage;
 
     @BeforeMethod
     public void beforeTest() throws MalformedURLException {
-        configureAndroidDriver(false);
+        MutableCapabilities capabilities = new UiAutomator2Options();
+        driver = new AndroidDriver(new URL("http://127.0.0.1:4723/wd/hub"), capabilities);
         wikipediaPage = new WikipediaPage(driver);
     }
 
@@ -30,7 +35,8 @@ public class WikipediaTest extends BaseTest {
 
     @Test(dataProvider = "searchQueries")
     public void testWikipediaSearch(String searchQuery) {
-        wikipediaPage.search(searchQuery);
+        var results = wikipediaPage.search(searchQuery);
+        Assert.assertFalse(results.isEmpty());
     }
 
     @DataProvider
