+++ modulo-4/mobileAutomation/src/test/java/tests/YouTubeTest.java
@@ -13,7 +13,7 @@ public class YouTubeTest extends BaseTest {
 
     @BeforeMethod
     public void beforeTest() throws MalformedURLException {
-        configureAndroidDriver(true);
+        configureAndroidDriver();
         youtubePage = new YouTubePage(driver);
     }
