+++ modulo-4/mobileAutomation/src/test/resources/cloud-suite.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
+<suite name="Cloud BrowserStack Tests">
+    <test name="Wikipedia Mobile Automation">
+        <classes>
+            <class name="tests.WikipediaTest"/>
+        </classes>
+    </test>
+</suite>
