--- modulo-4/youtubeAutomation/src/test/java/tests/BaseTest.java
@@ -1,45 +1,0 @@
-package tests;
-
-import io.appium.java_client.AppiumDriver;
-import io.appium.java_client.android.AndroidDriver;
-import io.appium.java_client.android.options.UiAutomator2Options;
-import io.appium.java_client.service.local.AppiumDriverLocalService;
-
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.time.Duration;
-
-public class BaseTest {
-
-    protected AppiumDriver driver;
-    protected AppiumDriverLocalService service;
-
-    protected void configureAndroidDriver(boolean isYoutube) throws MalformedURLException {
-        UiAutomator2Options options;
-        if (isYoutube) {
-             options = new UiAutomator2Options()
-                    .setDeviceName("emulator-5554")
-                    .setPlatformName("Android")
-                    .setAutomationName("UiAutomator2")
-                    .setAppPackage("com.google.android.youtube")
-                    .setAppActivity("com.google.android.youtube.HomeActivity")
-                    .amend("hideKeyboard",true)
-                    .setNewCommandTimeout(Duration.ofSeconds(60));
-        }
-        else {
-            options = new UiAutomator2Options()
-                    .setDeviceName("emulator-5554")
-                    .setPlatformName("Android")
-                    .setAutomationName("UiAutomator2")
-                    .setApp("./WikipediaSample.apk")
-                    .setAppPackage("org.wikipedia.alpha")
-                    .setAppActivity("org.wikipedia.main.MainActivity")
-                    .amend("hideKeyboard",true)
-                    .setNewCommandTimeout(Duration.ofSeconds(60));
-        }
-
-
-        driver = new AndroidDriver(new URL("http://127.0.0.1:4723"), options);
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(30));
-    }
-}
